Add API changelog
Fix import of get_service after rebase
